+++ module-services/service-audio/api/AudioServiceAPI.hpp
@@ -1,8 +1,9 @@
 #pragma once
 
 #include "../messages/AudioMessage.hpp"
-#include <optional>
 #include "Audio/decoder/decoder.hpp"
+
+#include <optional>
 
 class Service;